ReadBook_detail

+++ SpMVC_ReadBook/src/main/java/com/biz/book/mapper/BookDao.java
@@ -7,5 +7,5 @@ import com.biz.book.model.BookVO;
  * 기본 CRUD메서드를
  * 별도로 정의하지 않아도 된다.
  */
-public interface BookDao extends GenericDao<BookVO, Integer> {
+public interface BookDao extends GenericDao<BookVO, Long> {
 }
